Reverted imports.
I've reverted the imports from ItsSyfe commit due to imports on seperate line follows PEP-8 as it the script has more readability. Importing seperate functions instead of the entire package also applies to the point I made above about PEP-8 and can create conflicts if an object is defined in the script or another object has the same name.